apps/server/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.20
alembic>=1.12.0
celery>=5.3.4
redis>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.3.0
pydantic-settings>=2.0.3
python-multipart>=0.0.6
httpx>=0.24.1
pymysql>=1.1.0
cryptography>=41.0.4
pytest>=7.4.2 